move MockTableEngine into test-util
