+++ src/table/src/test_util.rs
@@ -1,5 +1,7 @@
 mod empty_table;
 mod memtable;
+mod mock_engine;
 
 pub use empty_table::EmptyTable;
 pub use memtable::MemTable;
+pub use mock_engine::MockTableEngine;
